move enable_tendermint_without_balance/enable_bch_with_tokens_without_balance to where it's used
