+++ mm2src/mm2_test_helpers/src/for_tests.rs
@@ -1736,41 +1736,6 @@ pub async fn enable_bch_with_tokens(
     json::from_str(&enable.1).unwrap()
 }
 
-pub async fn enable_bch_with_tokens_without_balance(
-    mm: &MarketMakerIt,
-    platform_coin: &str,
-    tokens: &[&str],
-    mode: UtxoRpcMode,
-    tx_history: bool,
-) -> Json {
-    let slp_requests: Vec<_> = tokens.iter().map(|ticker| json!({ "ticker": ticker })).collect();
-
-    let enable = mm
-        .rpc(&json!({
-            "userpass": mm.userpass,
-            "method": "enable_bch_with_tokens",
-            "mmrpc": "2.0",
-            "params": {
-                "ticker": platform_coin,
-                "allow_slp_unsafe_conf": true,
-                "bchd_urls": [],
-                "mode": mode,
-                "tx_history": tx_history,
-                "slp_tokens_requests": slp_requests,
-                "get_balances": false,
-            }
-        }))
-        .await
-        .unwrap();
-    assert_eq!(
-        enable.0,
-        StatusCode::OK,
-        "'enable_bch_with_tokens' failed: {}",
-        enable.1
-    );
-    json::from_str(&enable.1).unwrap()
-}
-
 pub async fn enable_solana_with_tokens(
     mm: &MarketMakerIt,
     platform_coin: &str,
@@ -2528,43 +2493,6 @@ pub async fn enable_tendermint(
             "tokens_params": ibc_requests,
             "rpc_urls": rpc_urls,
             "tx_history": tx_history
-        }
-    });
-    println!(
-        "enable_tendermint_with_assets request {}",
-        json::to_string(&request).unwrap()
-    );
-
-    let request = mm.rpc(&request).await.unwrap();
-    assert_eq!(
-        request.0,
-        StatusCode::OK,
-        "'enable_tendermint_with_assets' failed: {}",
-        request.1
-    );
-    println!("enable_tendermint_with_assets response {}", request.1);
-    json::from_str(&request.1).unwrap()
-}
-
-pub async fn enable_tendermint_without_balance(
-    mm: &MarketMakerIt,
-    coin: &str,
-    ibc_assets: &[&str],
-    rpc_urls: &[&str],
-    tx_history: bool,
-) -> Json {
-    let ibc_requests: Vec<_> = ibc_assets.iter().map(|ticker| json!({ "ticker": ticker })).collect();
-
-    let request = json!({
-        "userpass": mm.userpass,
-        "method": "enable_tendermint_with_assets",
-        "mmrpc": "2.0",
-        "params": {
-            "ticker": coin,
-            "tokens_params": ibc_requests,
-            "rpc_urls": rpc_urls,
-            "tx_history": tx_history,
-            "get_balances": false
         }
     });
     println!(
